+++ RenderSystems/Direct3D9/include/OgreD3D9RenderSystem.h
@@ -182,7 +182,7 @@ namespace Ogre
 		void convertColourValue( const ColourValue& colour, uint32* pDest );
 		void setStencilCheckEnabled(bool enabled);
         void setStencilBufferParams(CompareFunction func = CMPF_ALWAYS_PASS, 
-            ulong refValue = 0, ulong mask = 0xFFFFFFFF, 
+            uint32 refValue = 0, uint32 mask = 0xFFFFFFFF, 
             StencilOperation stencilFailOp = SOP_KEEP, 
             StencilOperation depthFailOp = SOP_KEEP,
             StencilOperation passOp = SOP_KEEP,
